chore: move withScrolling into utils, export utils from package
