+++ src/packages/solid/components/Column/Column.tsx
@@ -2,7 +2,7 @@ import { type Component } from 'solid-js';
 import { type IntrinsicNodeProps } from '@lightningjs/solid';
 import { Column as SolidColumn } from '@lightningjs/solid-primitives';
 import styles from './Column.styles.js';
-import { withScrolling } from '../../withScrolling.js';
+import { withScrolling } from '../../utils/withScrolling.js';
 
 export interface ColumnProps extends IntrinsicNodeProps {
   /** Item index at which scrolling begins */
