+++ src/packages/solid/components/Row/Row.tsx
@@ -2,7 +2,7 @@ import { type Component } from 'solid-js';
 import { type IntrinsicNodeProps } from '@lightningjs/solid';
 import { Row as SolidRow } from '@lightningjs/solid-primitives';
 import styles from './Row.styles.js';
-import { withScrolling } from '../../withScrolling.js';
+import { withScrolling } from '../../utils/withScrolling.js';
 
 export interface RowProps extends IntrinsicNodeProps {
   scrollIndex?: number;
