+++ src/packages/solid/utils/index.ts
@@ -1,2 +1,3 @@
 export { default as mapToneToStateHook } from './mapToneToStateHook.js';
 export { makeComponentStyles } from './getThemeStyles.js';
+export { withScrolling } from './withScrolling.js';
